Give library DLLs names of the form "vtLIBNAME.dll" on Cygwin.
This is because the Cygwin linker looks for LIBNAME.lib before
LIBNAME.dll, and we have LIBNAME.lib files around for the occam linker;
there doesn't seem to be an obvious way of turning off that behaviour,
so this is an easy workaround. It also significantly simplies the kroc
script's handling of libraries on Cygwin, since it doesn't need to hunt
for their paths by hand any more.