Adjust another overly agressive timer value on cgtest08.
The main problem is that on *BSD based operating systems the OS typically deschedules a process for a whole scheduling quantum when we may only wish to sleep for a few tens of microseconds.

The workaround is to "export CCSP_MINIMUM_SLEEP=10000", to spin instead of sleeping for small timer values (infact Mac OS defaults to 2000), but the downside of such a default would be wasted CPU time on applications where the timing is not (so) important.