"Port" CCSP and libkrocif to MinGW.
This is the nasty part of the port at the moment, because it essentially
consists of disabling the chunks of CCSP and libkrocif that rely on Unix
features (notably signals and pipes). I've done a quick hack around the
ccsp_safe_pause stuff for timers in CCSP, but blocking systems calls are still
broken, and as a result example programs can't do keyboard input properly.

Refs #116.