I'm unconvinced about this changeset in terms of performance, but I'll push it out so I can look at the buildbot results.  This changes the return semantics so that we allow the compiler to unwind in all but the cases where we return to a different process, for some small kernel calls (getpri/etc) this generates amazingly small call stubs, it also saves use setting up EBP/ESI for the return as the compiler saves these on the stack.  Generally even more assembly is deleted.
Additionally the return mechanism for calls with multiple parameters in the return has changed on IA32, only the first parameter is returned in a register (EAX), the rest are return in the scheduler call parameter slots.  This is a change I intend to keep as it fits with the entry system.  Again, everything is hidden abstraction macros.