Get mobile arrays working.
The only caveat here is that we can't fill in the mobile type properly, since
the old type descriptor system doesn't have as much information as the new one
-- we therefore allocate everything as arrays of bytes, and then set the
dimensions explicitly afterwards. On the upside, when the array we're receiving
is 1D, we can just reuse the mobile that the pony kernel sent us in the same
way.

Refs #6.