Fix DEFINED on dynamic mobiles.
This might not be the best fix, in that it doesn't change DEFINED; it just
makes sure that the field that DEFINED is looking at (which is the same field
that CHECKNOTNULL etc. use, i.e. what loadmobile loads) is actually cleared
properly when making a mobile undefined. Previously it wasn't initialised at
all.

This also changes the rather dubious-looking special code for clearing a
dynamic mobile array to use genmobileundefine like all the other mobile types,
so that a mobile array is undefined after it's been communicated away.

I've added a series of tests for DEFINED to cgtest85; it wasn't previously
exercised at all in the cgtests.