Provide a switch for turning the camera's auto-gain and auto-exposure on and off, under the name of auto.adjust.  Automatic white balancing is left on, and trust me you don't really want to switch it off...
The (underneath) implementation of this will definitely change, as I'm going to push the camera I2C configuration into the occam firmware, rather than do it in C.  However, this is spending on me putting together the I2C interface for the firmware (should be fairly trivial).

For now the idea is that you turn auto.adjust on for a while (couple of seconds), then turn it off and the camera will be adjusted for the lighting conditions of the present arena.