Interrupt driven UART0 output.  We still spin on the UART buffer, but when the MatchPort becomes full we hand off the remaining work to a low-priority interrupt.  Whenever the MatchPort next becomes ready we start pushing it data straight away from the interrupt.
Overall this change means we can in theory be getting data from the camera, encoding a frame to JPEG and pushing data out the UART "at the same time".  It also means that if the MatchPort stalls us because there is no one listening on the TCP socket then the whole TVM doesn't become blocked, we can gracefully block on the MatchPort from within occam and even allow the CPU to go to sleep.