Clean-up CIF stub generation; fixing stubs for non-rescheduled code.
Also optimise the stubs by playing off on the fact that the scheduler and the stack pointer are now the same following #127.  This optimisation is something I haven't done for tranx86 generated code yet.  Overall this saves some loads and stores on kernel calls.  There is probably still a bit of scope for optimising the rescheduled calls, but this code works well at the moment.

Fixes #149