Make FPU_NONE -1 rather than 0.
I think this fixes a subtle bug: since 0 is also the value of FPU_N (round to
nearest), if the FPU happened to be in FPU_N at the end of the previous PROC,
it wouldn't be reset to FPU_N at the start of the next one. -1 isn't a valid
rounding mode, so it'll cause it to always be reinitialised correctly.