A slight bit of reorganization of the UART code. Make firmware run.
- fluke/fluke_uart.c
- fluke/tvm_uart.c
- fluke/fluke_uart.h
- fluke/tvm_uart.h
These files were renamed and moved around simply to allow us to (for now) use
UART code from both projects (the Fluke and Surveyor port). Sections #ifdef'd
out where interrupt-driven behavior was required. 

The "in" portion of our UART code seems to crash the VM, so that will require
some investigation. This could be due to the fact that we have not implemented
sleep, etc.

- fluke/fluke.h
Included the UART headers here.

- fluke/main.c 
Used the fluke header instead of the UART header.

-- fluke/debug.c
Same.

--fluke/firmware/firmware.occ
--fluke/firmware/out.occ
Firmware now does "something". In particular, it prints to the screen and loops
forever. Included the "out" routines from the SRV1.

--fluke/Makefile
Added code to make the occam firmware, and a dependency to make sure that
happens when we make changes to the firmware.

Need to remove the .h firmware file at this point. Next commit.