This is an interim commit. I discovered an example regarding interrupt-driven code on the LPC series by R O Software. Parts of code from this source are already included in the Fluke distribution, but with constants renamed or tweaked. I've imported the example wholesale into the directory "lpc".
What this does is gives me a completely self-consistent set of headers in one
place, that completely specify connection to the board. I've moved everything to
use the new headers (I think), and will now continue to trim and comment out
things in the "original" Fluke code that had used this in a piecemeal fashion.
The end-result should be a more "textbook" port to the LPC2106, or, at least, a
port that is in keeping with widely available, freely available example code
that is used by many and pointed to from a number of fora.