This completes our exploration of the new TVM scheduler. There are choices we can make when deciding whether a channel is ready (using either the data or the channel word---choosing which is "volatile", for example).
We have a blinkenlight. A very complex blinkenlight.

On the todo list are things like idling (so we can do AFTER in a sane way),
making our interrupt code a bit more generic (so it might be used in a
cross-platform manner), cleaning up all of our code (we've hacked together some
ugly from all over the SRV1 wrapper), linking in UARTs (now that we have a
rudimentary understanding of the scheduler), tying in the Bluetooth, setting up
user contexts...

We're getting there.