I'm making a big ugly commit which consists of the exploration of some strange behaviour we observed when waiting on a timer list (tim ? AFTER ...) with respect to the VM busy waiting (which it should not) instead of entering idle/powerdown mode.
I'll clean this up in the next commit, this still run, there is just a lot of
garbage, and some of the implementation is not quite right.

What can be observed in first in the Fluke firmware (and later seen in
the SRV1) is that when both were waiting on the UART in the firmware for
characters (ie nothing on the timer queue) both firmwares went into
powerdown/idle mode. However, I tested things with timers initially in on the
Fluke and discovered that the it was busy waiting, as tvm_interrupt_pending()
in go_to_sleep() in timer.c always returned a non-zero value. I tested this in
the SRV1 firmware as well by inserting a process which put itself on the timer
list (tim ? AFTER ...) and observed that it didn't busywait either.

It turns out that the runtime sets a flag in sflags when there is something on
the timer queue, and the implementation of tvm_interrupt_pending that we
thieved from the SRV1 firmware returned the ORed value of the user and firmware
sflags, w/o masking out the VM specific ones.

A better tvm_interrupt_pending would probably be:
  #define INTERRUPT_SFLAGS  ((0xFFFFFFFF << SFLAG_USER_P) | SFLAG_INTR)
  return ((firmware_ctx.sflags & INTERRUPT_SFLAGS) |
          (user_ctx.sflags & INTERRUPT_SFLAGS));
Or using the TVM_INTR_SFLAGS that I just found:
  return ((firmware_ctx.sflags & TVM_INTR_SFLAGS) |
          (user_ctx.sflags & TVM_INTR_SFLAGS));
Or the way that SFLAG_INTR is being used in raise_tvm_interrupt, assuming that
it only ever is set when a HW interrupt is also pending, this would work too:
  return ((firmware_ctx.sflags & SFLAG_INTR) |
          (user_ctx.sflags & SFLAG_INTR));

Now for a bit of a cleanup.