Added the SConscripts for mkoccdeps and libtvm.
The mkoccdeps script is complete; it builds and cleans.

The libtvm script is a stub; we created it, and then started with mkoccdeps, because it had fewer things that needed to be known to make it compile... meaning, fewer #defines.

Updated the top-level SConstruct; minor changes, comments. Updated the .gitignore to include build output and auto-generated files like "config.h".