Create a branch for experimenting with automake.
I want to have a quick play with cleaning up KRoC's existing build
system, with the aim of getting rid of the build script and just having
everything be one large automake setup. I still think moving to a
different build system is a better long-term goal (not least because
automake's support for cross-directory dependencies is pretty awful),
but this should help with that anyway.

Of course, this may fail horribly, in which case this branch'll
disappear quickly...