Make it possible to build modules conditionally.
With the build script, it was OK for a module to fail to configure. This
isn't the case any more: when a module has prerequisites that aren't
there, the configure script must complete and the Makefile.am must then
not build anything. This adds an argument to OCCAM_PROVIDE so that
modules can be conditionally enabled; if a module is disabled, it will
be removed from in-tree-modules so that later configure scripts won't
find it.

This gets the remaining three KRoC modules added to the build: pony is
built if we're building with a pony-enabled KRoC, video is built if its
libraries are present, and udc is built only if it's specifically
enabled ('cause it's broken). The only module left that's not included
in the build is nocclibs, and that's because it's a bit odd; occbuild
doesn't yet have support for building whatever it's trying to build.