Fix a bug where a Trap "outbound" process could manage to register itself twice with its selector under heavy load.
I think (second guess...) this was what was stopping Occoids from
working well on the Display Wall: tracing the simulation running showed
that when it jammed, a ghost process had sent a Trap message that the
corresponding spirit never received. With this change it appears to work
happily, even with very large numbers of boids.