Fix the autoreconf-not-detecting-subdirs problem.
Sleeping on the problem helped: the solution is to actually call
AC_CONFIG_SUBDIRS in our equivalent macro (so autom4te detects it's
being called), but undo its effects that we don't want. This also avoids
duplicating some of AC_CONFIG_SUBDIRS' logic in
OCCAM_CONFIG_SUBDIRS_FOR, which is nice.