Do away with the last of the Arduino libraries in the Arduino wrapper. This only saves us about 600 bytes now, since we weren't using very much of it anyway.
I've imported the timer setup and reading code from the Arduino source,
and simplified it a bit since there's no way this code'll ever run on
the ATmega8. time_micros isn't currently being used, but we'll bind it
in the future.

I've rewritten pinMode and digitalWrite in pure occam -- mostly so
blink.occ will work, since I expect us to do something smarter once we
wrap IO up in processes. This means wait.for.interrupt is the only FFI
call at the moment.