Adding branch for CHAN OF SIGNAL
Christian and I spent a day exploring occ21. In particular, we wanted to
explore the feasibility of supporting the following idiom:


CHAN SIGNAL s:
PAR
  s ! SIGNAL
  s ? SIGNAL

This branch will hold the results of our exploration so that it can be seen and
considered by others. The subsequent commits will include our changes---this is
a clean commit of the branch.