occ21 modified to support CHAN SIGNAL
Our first commit in this series introduces support for channels with the
protocol SIGNAL. This makes the code

CHAN SIGNAL s:
PAR
  s ! SIGNAL
  s ? SIGNAL
 
syntactically correct and "do the right thing" in occam-pi. 

This commit will be followed up by commits of tests for this language feature as
well as a discussion of our rationale for the exploration/language addition. A
short email will then be sent to pop-dev with a pointer to the longer discussion
and the branch itself.