Make VPATH builds complete.
VPATH builds did not previously complete due to the tools being unable to pick
up non-generated items included by modules. Ie .inc files. 

There were two problems:
  * --in-tree only pointing to the build directory
    (files such as /tvm/posix/tvmspecials.inc could not be found in VPATH build)
  * --search paths for modules only pointing to the build directory
    (files such as course's const.inc could not be found in VPATH build)

--in-tree is used, for example, to setup a path to the tvm-posix directory in
where there is a file called 'tvmspecials.inc'. However, when doing a VPATH
build this no longer points to the correct directory. Fix this by providing
also an --src-tree option to occbuild which sets up search paths pointing into
the source tree (as well as the paths set up pointing to the build tree by
--in-tree).

To fix search paths for modules, make OCCAM_PROVIDE not only drop the build
directory in which the file is built into the in-tree-modules file, but also
drop the source directory. This is picked up by OCCAM_INCLUDE, which sets a
search path to point to the source directory in addition to the existing search
path which points to the build directory.



Possible improvements.
--in-tree and --src-tree are not very symmetrical names. Perhaps --build-tree
and --src-tree or something along those lines would be better. Perhaps it is
not important.