This is a commit of work in progress to get buttons (enabling/disabling), the teriminal, and stopping of a program put back in, while trying to make things more modular.
Currently the terminal is now used again for desktop programs (though input is
not currently working). The running tasks are again put in a single variable in
occPlug and it is only possible to run one task at a time (this should be a
reasonable assumption). The quit button is still not wired in though.