Make a universial build of the TVM. Only occ21 and tvm need to be universal, plus the occam libraries need to be built for both BE and LE.
Currently it is possible to build and upload for the Arduino. It is possible to
compile programs for the TVM on PPC but it is not possible to execute them. As
the tvm fails with an error. A 'make check' build of kroc trunk on PPC fails
(the cgtests), so there is something that needs looking at there.