Use POSIX sigaction rather than signal, which makes the TVM mostly work on Solaris.
This is the standard portability problem with signal(2) -- it has dramatically
different semantics on SysV- and BSD-derived Unix systems. CCSP also uses it,
but it has #ifdefs for the two different behaviours; we could make the same
change there to simplify it a little.

MinGW doesn't have sigaction, but it doesn't have SIGALRM either, so it doesn't
matter.