Use sizeof (void *) to figure out the pointer size, rather than preprocessor checks.
While the comments talk about alignment, it's actually how big a pointer
is that it cares about, since the items in the freelists must be big
enough to store a pointer to the next item. Since the header only knew
about the Alpha as being 64-bit, this was breaking on x86-64, with occ21
blowing up in the Occade examples owing to corrupt pointers in the
freelists.

I think this is an Inmos bug, since on machines with no alignment
requirement it would have been allocating 1-byte chunks then storing
pointers into them. I guess this'd usually have been masked by the
underlying malloc rounding up...