Trying to get some things to build for Windows.
This adds a bunch of old code I had hanging around for building packages. This
might not stay I just just wanted to avoid rewriting a bunch of code... I have
building of one package nearly working now, which I did as a test to see if I
could drive the build from python on windows, however:

Turns out that building in msys with a native windows python is not ideal. msys
deals in UNIX style paths, and native python does not. Bridging between the two
is not nice. Windows python does for example not know what to do with a #! line
so the script has to read that and do the right thing (ick). In spite of these
issues I have got things nearly working, but not quite, there is currently some
issues with \ vs. /. I have not yet tried to fix this and it may or may not be
possible. In moving towards getting windows stuff building here are some
options:

1) Try harder to get things to build using the setup i'm working with now
   - May not be possible...
2) Build in cygwin
   - I don't like cygwin and especially cyglib.dll or whatever it is called
3) Crossbuild on a non windows machine
   - Donno how much effort this would be
   - Upside is not having to deal with windows at all
   - Creating windows packages might be harder?
   - Wine might be helpful?
4) Don't try to script the build using python
   - Use shell scripting instead
   - However, occbuild is python, so it does not get rid of the python 
     requirement on windows
5) Build using Microsoft compilers
   - No!