[MCJ] We're seeing strange behavior, but committing regardless. (We want to take a break, so are committing the strangess, and will spend some time reading instead of poking.)
Specifically, we're having a hard time with PWM on pins 11 and 3. When we run main() in pwm-smoke-test we get all of the pins running PWM. When we run just pin 3, we don't see anything, and when we run just pin 11, it works. Clearly, our PWM init is wrong for TIMER2, we suspect.

Hence, we'll walk away, have dinner, revisit the documentation for the ATmega328, and see if we can figure out where we went wrong. (Are we running into the fact that the VM is using this timer already?)

Anyway. It's been a long enough afternoon, we'll come back to it.