A tri-fade example with analog input
Assuming LEDs on three PWM pins (3, 6, and 11 in this case) and a linear slide
potentiometer on pin 2 (arbitrary), this fades three LEDs out-of-sync. At the
end of a PWM fade cycle, a new reading is taken from the ADC, and that is used
to calculate the next fade rate.

I'd appreciate feedback on this code. It seems like it is unnecessarily complex
considering what I am tryng to do. My replicated ALT might just be me providing
protection for the critical section in the adc() process. Hence, it might not
be possible to simplify this code at this time. If that is the case, then we'll
leave this for my students to rediscover (or study) later.

Regardless, comments welcome.