Converted interrupt-driven to busywait
I converted adc() to a busywait version of the same PROC. I did not disable the
interrupts (if that needed to be done...), but instead simply wait until
conversion is complete in the runtime. 

While this is blocking, it allows me to do ADC from multiple pins. Before,
parallel ADC would deschedule while waiting for the interrupt, allowing
multiple processes to start conversions at the same time. I commented out the
previous solution for now, and will revisit this when I have time to read the
documentation on the ADC more closely.

I have tested this minimally (test program to follow), and will have to see if
everything else continues to be happy...

Tests for the Wiring and Plumbing layers might be nice.