Make AVR tools for installation in /opt
Instead of leaving this in a personal repository, I've decided to put it under "distribution". Really, it may have been better just to make a Debian package...

This checks out a clean repository into

~/svn/avr-...

does a cross-build, and installs it into

~/pkg/avr/...

it then creates a tarball that expands to a directory structure rooted in

occam/arduino

which is intended to live under /opt.

This may be of use to others. Really, it is only intended to support the use of our AVR tools with students at Allegheny for the short term... that is, until I do a proper packaging for various *NIXes. However, it taught me useful things, so that should make creating other packages relatively easy... or, easier, anyway.