Make occbuild work with Python 2.3 again.
It appears I might have spoken a bit soon about being able to rely on 2.4+
(i.e. the subprocess module) -- York's grid is using an older version of
Scientific Linux which ships with nothing newer that 2.3, so we need to
continue to support that for now. This patch tries subprocess first, and then
falls back to os if that's not available, so it'll work on Windows too.