Branch for servo support exploration.
I'm not sure if this is the "right" way to go. We could tag the trunk at stable
points, and let it remain unstable... or, we could branch, and explore there.

For the moment, I'm creating a branch for Dennis to commit his servo work. If I
need to do something to prep things for merging at this point, someone should
let me know. I'm not a merge master, I'm afraid to say...