Branch for implementing TWI
Again, perhaps I should just do this in trunk, as the changes to the runtime
are minimally invasive...

So, I'll continue my branch-fest, and commit my work from the weekend (that
doesn't seem to work yet) on implementing I2C/TWI here. It will be instructive
to have things in the repos anyway... this is my first foray into implementing
a protocol/state machine in o-pi, so comments/feedback are going to be very
welcome.