Comments on community code.
This has been too long coming. I don't know if feedback was being waited on or not... if it was, we weren't responsive enough as a group. If there was feedback expected, and we weren't responsive... someone should have harassed us!

Anyway, there are some comments on some of the code that bringfire contributed. I like the design; I made a few small revisions, and tried to comment my additions so that my rationale would be in-line.

Questions/comments welcome.