Very crude Deb binary pkg builder.
This script, when executed, checks out the source tree, builds it for AVR, and creates a (gross) Debian package. Much to be tweaked/improved... but, it creates *something* that works at the moment.

My longer-term goal is to have a Debian source package that we can upload to Launchpad as a PPA. If we get there before OSCON, that's great. Otherwise, a rough-and-tumble .pkg will have to do.

I may get around to creating a repository structure and making that part of the automation, but I think there's a server move coming up, so I'm hesitant to do that right now.