Adding a "better" build script.
I say "better" because these are both kinda rough. This breaks out more of the build stages (easier to test the build script). Note that the file I just deleted should probably have been kept, and a 'sed' performed over it, instead of echoing it from the script.

These things can be improved in time, depending on how long this script lives.

I'll be removing the old script and renaming this one. I know that isn't the nicest way to do things in the repos, but whats done is done... and this isn't exactly critical path stuff.